+++ library/OverlayElementButtonClickEvent.mjs
@@ -1,0 +1,5 @@
+/** @typedef {import("./OverlayElementButtonClickEventDetail.mjs").OverlayElementButtonClickEventDetail} OverlayElementButtonClickEventDetail */
+
+/**
+ * @typedef {CustomEvent<OverlayElementButtonClickEventDetail>} OverlayElementButtonClickEvent
+ */