+++ library/OverlayElementInputChangeEvent.mjs
@@ -1,0 +1,5 @@
+/** @typedef {import("./OverlayElementInputChangeEventDetail.mjs").OverlayElementInputChangeEventDetail} OverlayElementInputChangeEventDetail */
+
+/**
+ * @typedef {CustomEvent<OverlayElementInputChangeEventDetail>} OverlayElementInputChangeEvent
+ */