+++ library/OverlayElementInputChangeEventDetail.mjs
@@ -1,0 +1,5 @@
+/** @typedef {import("form/FormElementChangeEventDetail.mjs").FormElementChangeEventDetail} FormElementChangeEventDetail */
+
+/**
+ * @typedef {FormElementChangeEventDetail} OverlayElementInputChangeEventDetail
+ */