+++ library/OverlayElementInputInputEvent.mjs
@@ -1,0 +1,5 @@
+/** @typedef {import("./OverlayElementInputInputEventDetail.mjs").OverlayElementInputInputEventDetail} OverlayElementInputInputEventDetail */
+
+/**
+ * @typedef {CustomEvent<OverlayElementInputInputEventDetail>} OverlayElementInputInputEvent
+ */