+++ library/OverlayElementInputInputEventDetail.mjs
@@ -1,0 +1,5 @@
+/** @typedef {import("form/FormElementInputEventDetail.mjs").FormElementInputEventDetail} FormElementInputEventDetail */
+
+/**
+ * @typedef {FormElementInputEventDetail} OverlayElementInputInputEventDetail
+ */